local-multicast/pyproject.toml
[tool.poetry]
name = "local-multicast"
version = "0.1.0"
description = ""
authors = ["Reed Trevelyan <reedtrevelyan@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
PyYAML = "^5.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
